+++ src/realm-settings/key-providers/rsa-generated/RSAGeneratedForm.tsx
@@ -1,12 +1,7 @@
 import React, { useState } from "react";
 import {
-  AlertVariant,
-  Button,
-  ButtonVariant,
   Form,
   FormGroup,
-  Modal,
-  ModalVariant,
   Select,
   SelectOption,
   SelectVariant,
@@ -15,24 +10,19 @@ import {
 } from "@patternfly/react-core";
 import { useTranslation } from "react-i18next";
 import { Controller, useForm } from "react-hook-form";
-
-import { useAdminClient } from "../context/auth/AdminClient";
-import { useAlerts } from "../components/alert/Alerts";
 import type ComponentRepresentation from "keycloak-admin/lib/defs/componentRepresentation";
-import { HelpItem } from "../components/help-enabler/HelpItem";
-import { useServerInfo } from "../context/server-info/ServerInfoProvider";
-import { useRealm } from "../context/realm-context/RealmContext";
+import { HelpItem } from "../../../components/help-enabler/HelpItem";
+import { useServerInfo } from "../../../context/server-info/ServerInfoProvider";
 
 type RSAGeneratedFormProps = {
   providerType?: string;
   handleModalToggle?: () => void;
   refresh?: () => void;
-  save: (component: ComponentRepresentation) => void;
+  save?: (component: ComponentRepresentation) => void;
 };
 
 export const RSAGeneratedForm = ({
   providerType,
-  handleModalToggle,
   save,
 }: RSAGeneratedFormProps) => {
   const { t } = useTranslation("groups");
@@ -236,4 +226,13 @@ export const RSAGeneratedForm = ({
         </>
     </Form>
   );
+};
+
+export const RSASettings = () => {
+
+  return (
+     <RSAGeneratedForm />
+  );
+
+
 };
